pyproject.toml
[project]
name = "2d-metaverse"
version = "0.1.0"
description = ""
authors = [
    {name = "skyspec28",email = "maulomepelumi@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "3.13.1"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "selenium (>=4.28.1,<5.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "django-allauth[socialaccount] (>=65.3.1,<66.0.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "markdown (>=3.7,<4.0)",
    "django-filter (>=24.3,<25.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
